Cleaning dispatcher file structure.
