+++ Makefile
@@ -31,7 +31,8 @@ MODULES= \
 	src/config/config.o \
 	src/io_management/workspace_settings.o \
 	src/advanced/lsp/lsp_handler.o \
-	src/advanced/lsp/lsp_diagnostic_handler.o \
+	src/advanced/lsp/lsp_notification_dispatcher.o \
+	src/advanced/lsp/lsp_response_dispatcher.o \
 	src/advanced/lsp/lsp_highlighter.o \
 	src/advanced/lsp/lsp_dispatcher.o \
 	src/advanced/lsp/lsp_emitter.o \
