--- src/advanced/lsp/lsp_diagnostic_handler.c
@@ -1,4 +1,0 @@
-#include "lsp_diagnostic_handler.h"
-
-
-void receiveDiagnostic(cJSON* packet, long* payload) {}
