+++ src/advanced/lsp/lsp_dispatcher.h
@@ -13,4 +13,10 @@ typedef struct {
 
 void dispatcher(cJSON* packet, LSP_Server* lsp, void* payload);
 
+int getIndexFileContainerForUri(DispatcherPayload* payload, cJSON* params);
+
+int getIndexFileContainerForName(DispatcherPayload* payload, char* file_name);
+
+void printPacket(cJSON* packet, cJSON* params);
+
 #endif // WISHWIM_LSP_DISPATCHER_H
