+++ src/advanced/lsp/lsp_notification_dispatcher.h
@@ -2,8 +2,9 @@
 #define WISHWIM_LSP_DIAGNOSTIC_HANDLER_H
 
 #include "../../../lib/cJSON/cJSON.h"
+#include "lsp_dispatcher.h"
 
-void receiveDiagnostic(cJSON* packet, long* payload);
 
+void notificationDispatcher(cJSON* packet, DispatcherPayload* data);
 
 #endif // WISHWIM_LSP_DIAGNOSTIC_HANDLER_H
