+++ src/advanced/lsp/lsp_response_dispatcher.h
@@ -1,0 +1,7 @@
+#ifndef WISHWIM_LSP_RESPOSNE_DISPATCHER_H
+#define WISHWIM_LSP_RESPOSNE_DISPATCHER_H
+#include "lsp_dispatcher.h"
+
+void responseDispatcher(cJSON* packet, LSP_Server* lsp, DispatcherPayload* data) ;
+
+#endif // WISHWIM_LSP_RESPOSNE_DISPATCHER_H
